refactor: moved file to rulesTable
